+++ Assets/Resource/SlotIcon/BMJUA_TTF SDF.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 207b31966bab784448a6cc3d0ff9d6f4
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 0
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
