+++ Assets/Resource/SlotIcon/ShopItem.txt
@@ -1,0 +1,1 @@
+[{"type":"Equip","name":"Helmet","price":3000,"Hp_up":300.0,"Movespeed_up":0.0,"FireRate_up":0.0},{"type":"Food","name":"Meat","price":500,"Value":350.0},{"type":"UsableItem","name":"Axe","price":650,"Battle_type":0,"Durability":200,"Damage":50,"PerUsability":5,"Accuarancy":100.0,"Attack_speed":10.0,"Value":0.0},{"type":"UsableItem","name":"Bow","price":70,"Battle_type":1,"Durability":5,"Damage":30,"PerUsability":3,"Accuarancy":60.0,"Attack_speed":5.0,"Value":0.0},{"type":"UsableItem","name":"Steroid","price":3000,"Battle_type":2,"Durability":0,"Damage":0,"PerUsability":0,"Accuarancy":0.0,"Attack_speed":0.0,"Value":50.0},{"type":"UnUsable","name":"Gem","price":5000}]
