+++ Assets/Resource/SlotIcon/ShopItem.txt.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8425d9615ef02c0408329cd65798e5ba
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
