+++ Assets/Scripts/ItemScripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: accbc926ce18a3a4585ddb76af5dc1ff
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
