+++ Assets/Scripts/ItemScripts/ShopManagement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f9a4d79ba6e9f9440876552e3e3c8dcb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
