--- Assets/Scripts/SlotScripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 1f1e7a3082a414949afb17e3b39d616c
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
