--- Assets/Scripts/SlotScripts/ItemDataInitialize.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ac1444734e04283419d0e9f241042601
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
