--- Assets/Scripts/SlotScripts/Popup.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 3643251f09f4c5c4eb9a4bb6ac6cb33d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
