+++ Assets/ShopPopUpManagement.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d47eb55fe0c34194598d522a4df2e044
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
